+++ player/src/components/videos/ReviewVideoPreview.tsx
@@ -7,6 +7,12 @@ type Props = {
 };
 
 const ReviewVideoPreview = observer(({ video }: Props) => {
+  const [videoNotStarted, setVideoNotStarted] = React.useState<boolean | null>(
+    null
+  );
+  const [videoFinished, setVideoFinished] = React.useState<boolean | null>(
+    null
+  );
   const containerEl = React.useRef<null | HTMLDivElement>(null);
   const selectedVideo = video.session.selectedVideo;
 
@@ -18,22 +24,62 @@ const ReviewVideoPreview = observer(({ video }: Props) => {
     if (
       containerEl.current === null ||
       video.reviewVideoEl === null ||
-      (selectedVideo !== null && selectedVideo.id === video.id)
+      (selectedVideo !== null && selectedVideo.id === video.id) ||
+      videoNotStarted === true ||
+      videoFinished === true
     ) {
       return;
     }
 
     video.reviewVideoEl.volume = 0;
     containerEl.current.appendChild(video.reviewVideoEl);
-  }, [video.videoElementsCreated, selectedVideo]);
+  }, [
+    video.videoElementsCreated,
+    selectedVideo,
+    videoNotStarted,
+    videoFinished,
+  ]);
+
+  React.useEffect(() => {
+    if (
+      video.session.currentTime == null ||
+      video.beginsAt === null ||
+      video.finishesAt === null
+    ) {
+      return;
+    }
+
+    setVideoNotStarted(video.session.currentTime < video.beginsAt);
+    setVideoFinished(video.session.currentTime > video.finishesAt);
+  }, [video.session.currentTime, video.beginsAt, video.finishesAt]);
 
   const style = { aspectRatio: `${video.width}/${video.height}` };
 
   const renderedContent = (() => {
+    if (videoNotStarted === null) {
+      return null;
+    }
+
     if (selectedVideo !== null && selectedVideo.id === video.id) {
       return (
         <div className="flex items-center justify-center w-full italic text-white/50">
           Playing in main window
+        </div>
+      );
+    }
+
+    if (videoNotStarted === true) {
+      return (
+        <div className="flex items-center justify-center w-full italic text-white/50">
+          Before video starts
+        </div>
+      );
+    }
+
+    if (videoFinished === true) {
+      return (
+        <div className="flex items-center justify-center w-full italic text-white/50">
+          Video finished
         </div>
       );
     }