+++ player/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 26/02/2023
 
 - Added latest changes news
+- Timelines in the review video are now unified between all videos.
+- Preview videos now show when they have finished or not started.
 
 # 24/02/2023